+++ samples/lib/widgets/switch.dart
@@ -21,7 +21,7 @@ class _SwitchState extends StatefulSampleState<SwitchPage> {
   @override
   Widget sample(BuildContext context) => FSwitch(
     label: const Text('Airplane Mode'),
-    semanticLabel: 'Airplane Mode',
+    semanticsLabel: 'Airplane Mode',
     value: state,
     onChange: (value) => setState(() => state = value),
     enabled: widget.enabled,